Add badges to previews indicating profile changes
I initially planned to use a checksum of the current profile, filtered down to the relevant settings, using the json string as the ingest, but strangely enough the json exporter wasn't consistent (despite sort_keys=True), causing phantom changes to be reported.
I could've made a custom exporter, but that would've been another point needing updates every time something in the profile changed, so I took the lazy way out with dictdiffer, which was always accurate when it came to finding differences. Measuring it, it isn't that much slower, all still in the 0 to 10 ms range, but memory usage is obviously higher. I use caching already for these difference checks so the impact is minimized.