No longer scale pixel measures
With the change from a global scale factor to individual images being scaled to fit into a target range, it no longer makes sense to scale the profile options along with it. The global scale factor works best when you have a known size and configure your profile to work for that, then just use the scale to optimize the workload, while keeping all your settings proportional. But that typically doesn't happen. Instead you have a config optimized for one image size and suddenly encounter larger images. Instead of adjusting all of your parameters, you can now designate a target size your profile is meant for and large images are scaled down to work within it.