fix minor issues
I moved the OCR options a little further up to be with the rest of them.

More importantly, I added them into the actual string templates for the config, that's what lets the GUI see them too.

I also made the GUI sensitive to changes in those settings, see gui/image_file.py

in main.py:198 The way this works, is that docopt parses the --profile CLI option, and if there isn't one, it's set to None. The load_profile function detects this and only then uses the default profile in the config. So it's important to use the docopt args here, so that the CLI flag is actually respected.